// apps/mobile/http/generated/models/PutOrganizationsSlugCustomersCustomerid.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type PutOrganizationsSlugCustomersCustomeridPathParams = {
  /**
   * @type string
   */
  slug: string;
  /**
   * @type string
   */
  customerId: string;
};

export const putOrganizationsSlugCustomersCustomerid204Enum = {} as const;

export type PutOrganizationsSlugCustomersCustomerid204EnumKey =
  (typeof putOrganizationsSlugCustomersCustomerid204Enum)[keyof typeof putOrganizationsSlugCustomersCustomerid204Enum];

/**
 * @description Default Response
 */
export type PutOrganizationsSlugCustomersCustomerid204 =
  PutOrganizationsSlugCustomersCustomerid204EnumKey | null;

export const putOrganizationsSlugCustomersCustomeridMutationRequestPersonTypeEnum =
  {
    PF: "PF",
    PJ: "PJ",
  } as const;

export type PutOrganizationsSlugCustomersCustomeridMutationRequestPersonTypeEnumKey =
  (typeof putOrganizationsSlugCustomersCustomeridMutationRequestPersonTypeEnum)[keyof typeof putOrganizationsSlugCustomersCustomeridMutationRequestPersonTypeEnum];

export const putOrganizationsSlugCustomersCustomeridMutationRequestDocumentTypeEnum =
  {
    CPF: "CPF",
    CNPJ: "CNPJ",
    RG: "RG",
    IE: "IE",
    PASSPORT: "PASSPORT",
    OTHER: "OTHER",
  } as const;

export type PutOrganizationsSlugCustomersCustomeridMutationRequestDocumentTypeEnumKey =
  (typeof putOrganizationsSlugCustomersCustomeridMutationRequestDocumentTypeEnum)[keyof typeof putOrganizationsSlugCustomersCustomeridMutationRequestDocumentTypeEnum];

export const responsibleTypeEnum3 = {
  SELLER: "SELLER",
  PARTNER: "PARTNER",
} as const;

export type ResponsibleTypeEnum3Key =
  (typeof responsibleTypeEnum3)[keyof typeof responsibleTypeEnum3];

export type PutOrganizationsSlugCustomersCustomeridMutationRequest = {
  /**
   * @type string
   */
  name: string;
  /**
   * @type string
   */
  personType: PutOrganizationsSlugCustomersCustomeridMutationRequestPersonTypeEnumKey;
  /**
   * @type string | undefined
   */
  phone?: string;
  /**
   * @type string | undefined
   */
  email?: string;
  /**
   * @type string
   */
  documentType: PutOrganizationsSlugCustomersCustomeridMutationRequestDocumentTypeEnumKey;
  /**
   * @type string
   */
  documentNumber: string;
  /**
   * @type object
   */
  responsible?: {
    /**
     * @type string
     */
    type: ResponsibleTypeEnum3Key;
    /**
     * @type string, uuid
     */
    id: string;
  } | null;
  /**
   * @type object | undefined
   */
  pf?: {
    birthDate?: any;
    /**
     * @type number | undefined
     */
    monthlyIncome?: number;
    /**
     * @type string | undefined
     */
    profession?: string;
    /**
     * @type string | undefined
     */
    placeOfBirth?: string;
    /**
     * @type string | undefined
     */
    fatherName?: string;
    /**
     * @type string | undefined
     */
    motherName?: string;
    /**
     * @type string | undefined
     */
    naturality?: string;
  };
  /**
   * @type object | undefined
   */
  pj?: {
    /**
     * @type string | undefined
     */
    businessActivity?: string;
    /**
     * @type string | undefined
     */
    municipalRegistration?: string;
    /**
     * @type string | undefined
     */
    stateRegistration?: string;
    /**
     * @type string | undefined
     */
    legalName?: string;
    /**
     * @type string | undefined
     */
    tradeName?: string;
    foundationDate?: any;
  };
};

export type PutOrganizationsSlugCustomersCustomeridMutationResponse =
  PutOrganizationsSlugCustomersCustomerid204;

export type PutOrganizationsSlugCustomersCustomeridMutation = {
  Response: PutOrganizationsSlugCustomersCustomerid204;
  Request: PutOrganizationsSlugCustomersCustomeridMutationRequest;
  PathParams: PutOrganizationsSlugCustomersCustomeridPathParams;
  Errors: any;
};